Cache everything, including partials, lambdas & files
This makes several big changes. First, templates read from the
filesystem are cached as well as string literal templates.

Next, templates that include lambdas are cached. This did not require
any extra work, the precaution was unnecessary.

Next, partial templates are cached. This is the biggest change, and is
handled by integrating the 'get-template' and 'parse-template'
functions. Getting a template, no matter if it is done inside a partial
or at the top level, can pull from the cache or parse and store into the
cache.

The cache is now an instance member, rather than a state variable shared
among instances. This provides the user with more control, and should
conform better to the policy of least astonishment.

Small changes that got tangled up in this:
- extensions are now specified without a leading '.' (but '.foo' is
  still handled; it may be deprecated in the future)
- lambdas are not allowed to return the name of a partial; their return
  value is always treated as a template literal