Workaround for .hyper .map bugginess (t/11-iterable.t)
See Bug #41 for details. If $obj.Str is called when $obj is the
HyperSequence from t/11-iterable.t, then later on when .map is called it
dies:

    A worker in a parallel iteration (hyper or race) initiated here:
      in block <unit> at ./bug41.raku line 3

    Died at:
        Cannot invoke object with invocation handler in this context

When it is not stringified, it works OK.

This is brittle and probably needs a more permanent fix (in the test
case?), but for now this workaround helps.